Tests for form submissions. Needed some changes on the form component. But its all there now. Tests PDF generation, DCAD scraper, revisions, etc. Super happy with it. Wasn't sure if its gonna be doable.

+++ lib/sql/form/submission/get.sql
@@ -11,7 +11,7 @@ SELECT
   forms.name AS title,
   (
     SELECT count(*) FROM forms_data WHERE forms_data.submission = forms_submissions.id
-  ) as revision_count,
+  )::int as revision_count,
   'form_submission' AS type,
   (
     SELECT file FROM files_relations WHERE role = 'SubmissionRevision' AND role_id = forms_data.id
